# app/libs/http/controllers/product_controller/fiber_create.go
package product_controller

import (
	"net/http"
	"rabi-food-core/libs/database/gateways/product_gateway"
	"rabi-food-core/libs/validator"

	"github.com/gofiber/fiber/v2"
)

// CreateProduct godoc
// @Summary Create product
// @Description Create a new product
// @Tags products
// @Accept json
// @Produce text/plain
// @Param product body product_gateway.CreateInput true "Product data"
// @Success 201 {string} string "Created product ID"
// @Failure 400 {object} middlewares.ValidationErrorResponse "Validation errors"
// @Failure 500 {string} string "Internal server error"
// @Router /product/ [post]
func (c *ProductController) Create(ctx *fiber.Ctx) error {
	data := product_gateway.CreateInput{}
	err := ctx.BodyParser(&data)
	if err != nil {
		return ctx.JSON(err)
	}

	err = validator.V.Struct(data)
	if err != nil {
		return err
	}

	id, err := c.usecase.Create(ctx.UserContext(), data)

	if err != nil {
		return err
	}

	return ctx.Status(http.StatusCreated).SendString(id)
}